+++ CotEditor/Sources/CEEditorWrapper.m
@@ -592,11 +592,9 @@
 - (IBAction)toggleInvisibleChars:(nullable id)sender
 // ------------------------------------------------------
 {
-    BOOL showsInvisibles = ![(CELayoutManager *)[[self focusedTextView] layoutManager] showsInvisibles];
+    BOOL showsInvisibles = ![self showsInvisibles];
+    [self setShowsInvisibles:showsInvisibles];
     
-    [[self splitViewController] enumerateEditorViewsUsingBlock:^(CEEditorViewController * _Nonnull viewController) {
-        [viewController setShowsInvisibles:showsInvisibles];
-    }];
     [[[self windowController] toolbarController] toggleItemWithTag:CEToolbarShowInvisibleCharsItemTag
                                                              setOn:showsInvisibles];
 }